Use a single instance of the entity store to the webStorage.
The contentscript should not have anything to do with the webStorage
since the background page is the only handle for it. Redo the
Backbone.sync to communicate to the chrome API through messaging
so that all requests go to the same area.

Improved the data model from object to DDL for the abstract entity.